+++ api/dgram.js
@@ -16,7 +16,7 @@ import { InternalError } from './errors.js'
 import { EventEmitter } from './events.js'
 import diagnostics from './diagnostics.js'
 import { Buffer } from './buffer.js'
-import { isIPv4 } from './net.js'
+import { isIPv4 } from './ip.js'
 import process from './process.js'
 import console from './console.js'
 import ipc from './ipc.js'
